Refactor top menu into new file
